+++ packages/rfw/CHANGELOG.md
@@ -1,6 +1,9 @@
+## NEXT
+
+* Updates example app Gradle version to 7.6.3.
+
 ## 1.0.16
 
-* Updates Gradle version of example apps to 7.6.3.
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 * Fixes lint warnings.