+++ packages/rfw/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 * Fixes lint warnings.
+* Updates Gradle version of example apps to 7.6.3.
 
 ## 1.0.15