+++ packages/rfw/CHANGELOG.md
@@ -1,5 +1,5 @@
-## NEXT
-
+## 1.0.16
+* Adds a wasm web example
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 * Fixes lint warnings.